+++ com.dotsui.hybrid-samples/DotsUI.Samples/PrefabExample/PrefabExample.cs
@@ -11,7 +11,6 @@ using Unity.Mathematics;
 
 public class PrefabExample : MonoBehaviour
 {
-    [SerializeField] Canvas m_FpsCanvas;
 
     [SerializeField] Canvas m_TopCanvas;
     [SerializeField] Canvas m_RightCanvas;
@@ -22,21 +21,19 @@ public class PrefabExample : MonoBehaviour
     {
         if (World.Active == null)
             DefaultWorldInitialization.Initialize("UI World", false);
-        var entityManager = World.Active.EntityManager;
+        //var entityManager = World.Active.EntityManager;
 
         World.Active.GetOrCreateSystem<UserInputSystemGroup>().AddSystemToUpdateList(World.Active.GetOrCreateSystem<InstantiationSystem>());
         World.Active.GetOrCreateSystem<UserInputSystemGroup>().AddSystemToUpdateList(World.Active.GetOrCreateSystem<CloseButtonSystem>());
 
-        RectTransformToEntity transformToEntity = new RectTransformToEntity(100, Allocator.Temp);
-        RectTransformConversionUtils.ConvertCanvasHierarchy(m_FpsCanvas, World.Active.EntityManager, transformToEntity);
-        RectTransformConversionUtils.ConvertCanvasHierarchy(m_TopCanvas, World.Active.EntityManager, transformToEntity);
-        RectTransformConversionUtils.ConvertCanvasHierarchy(m_RightCanvas, World.Active.EntityManager, transformToEntity);
-        RectTransformConversionUtils.ConvertCanvasHierarchy(m_PrefabDestination, World.Active.EntityManager, transformToEntity);
+        //RectTransformToEntity transformToEntity = new RectTransformToEntity(100, Allocator.Temp);
+        //RectTransformConversionUtils.ConvertCanvasHierarchy(m_TopCanvas, World.Active.EntityManager, transformToEntity);
+        //RectTransformConversionUtils.ConvertCanvasHierarchy(m_RightCanvas, World.Active.EntityManager, transformToEntity);
+        //RectTransformConversionUtils.ConvertCanvasHierarchy(m_PrefabDestination, World.Active.EntityManager, transformToEntity);
 
-        GameObject.Destroy(m_FpsCanvas.gameObject);
-        GameObject.Destroy(m_TopCanvas.gameObject);
-        GameObject.Destroy(m_RightCanvas.gameObject);
-        GameObject.Destroy(m_PrefabDestination.gameObject);
-        transformToEntity.Dispose();
+        //GameObject.Destroy(m_TopCanvas.gameObject);
+        //GameObject.Destroy(m_RightCanvas.gameObject);
+        //GameObject.Destroy(m_PrefabDestination.gameObject);
+        //transformToEntity.Dispose();
     }
 }