+++ com.dotsui.hybrid-samples/DotsUI.Samples/PrefabExample/WindowCanvas.cs
@@ -4,12 +4,11 @@ using DotsUI.Hybrid;
 using Unity.Entities;
 using UnityEngine;
 
-public class WindowCanvas : MonoBehaviour, IRectTransformToEntity
+public class WindowCanvas : MonoBehaviour, IConvertGameObjectToEntity
 {
-    public void ConvertToEntity(Entity entity, RectTransformToEntity rectTransformToEntity, Dictionary<Object, Entity> assetToEntity,
-        EntityManager mgr)
+    public void Convert(Entity entity, EntityManager dstManager, GameObjectConversionSystem conversionSystem)
     {
-        mgr.AddComponent(entity, typeof(WindowCanvasComponent));
+        dstManager.AddComponent(entity, typeof(WindowCanvasComponent));
     }
 }