+++ com.dotsui.hybrid/DotsUI.Hybrid/Renderer/ScreenSpaceOverlaySystem.cs
@@ -27,38 +27,80 @@ namespace DotsUI.Hybrid.Renderer
     class ScreenSpaceOverlaySystem : JobComponentSystem
     {
         private List<CommandBuffer> m_Buffers = new List<CommandBuffer>();
-        private EntityQuery m_OverlayCanvasQuery;
+        private EntityQuery m_DirtyCanvasQuery;
 
         private GuiProxy m_Proxy;
+        private EntityQuery m_UnitializedCanvasQuery;
+        private EntityQuery m_ScreenSpaceOverlayQuery;
+        private EntityQuery m_DestroyedCanvasQuery;
+
+        struct OverlayCanvasInitialized : ISystemStateComponentData
+        {
+
+        }
         protected override void OnCreate()
         {
-            //Canvas.willRenderCanvases += OnAfterFrameRender;
-            m_OverlayCanvasQuery = GetEntityQuery(new EntityQueryDesc()
+            m_DirtyCanvasQuery = GetEntityQuery(new EntityQueryDesc()
             {
                 All = new ComponentType[]
                 {
                     ComponentType.ReadOnly<CanvasCommandBufferContainer>(),
-                    ComponentType.ReadOnly<CanvasMeshContainer>(),
                     ComponentType.ReadOnly<CanvasSortLayer>(),
-                    ComponentType.ReadOnly<RebuildCanvasHierarchyFlag>(),
+                    ComponentType.ReadOnly<RebuildCanvasHierarchyFlag>(), 
                     ComponentType.ReadOnly<CanvasScreenSpaceOverlay>(),
                 }
             });
-            m_Proxy = new UnityEngine.GameObject("ScreenSpaceGUIProxy_DO_NOT_DESTROY").AddComponent<GuiProxy>();
+            m_UnitializedCanvasQuery = GetEntityQuery(new EntityQueryDesc()
+            {
+                All = new ComponentType[]
+                {
+                    ComponentType.ReadOnly<CanvasCommandBufferContainer>(),
+                    ComponentType.ReadOnly<CanvasSortLayer>(),
+                    ComponentType.ReadOnly<CanvasScreenSpaceOverlay>(),
+                },
+                None = new ComponentType[]
+                {
+                    typeof(OverlayCanvasInitialized)
+                }
+            });
+            m_DestroyedCanvasQuery = GetEntityQuery(new EntityQueryDesc()
+            {
+                All = new ComponentType[]
+                {
+                    typeof(OverlayCanvasInitialized),
+                },
+                None = new ComponentType[]
+                {
+                    typeof(CanvasScreenSpaceOverlay),
+                }
+            });
+
+            m_ScreenSpaceOverlayQuery = EntityManager.CreateEntityQuery(new EntityQueryDesc()
+            {
+                All = new ComponentType[]
+                {
+                    ComponentType.ReadOnly<CanvasCommandBufferContainer>(),
+                    ComponentType.ReadOnly<CanvasSortLayer>(),
+                    ComponentType.ReadOnly<CanvasScreenSpaceOverlay>(),
+                }
+            });
+
+            RequireForUpdate(m_UnitializedCanvasQuery);
+            m_Proxy = new UnityEngine.GameObject("ScreenSpaceOverlayGUIProxy_DO_NOT_DESTROY").AddComponent<GuiProxy>();
             m_Proxy.onRenderGui = OnAfterFrameRender;
         }
 
         protected override void OnDestroy()
         {
             m_Buffers.Clear();
-            //GameObject.Destroy(m_Proxy.gameObject);
+            if(m_Proxy != null)
+                GameObject.Destroy(m_Proxy.gameObject);
         }
 
         private void OnAfterFrameRender()
         {
             for(int i = 0; i < m_Buffers.Count; i++)
             {
-                //RenderTexture.active = null;
                 Graphics.ExecuteCommandBuffer(m_Buffers[i]);
             }
         }
@@ -66,14 +108,31 @@ namespace DotsUI.Hybrid.Renderer
         protected override JobHandle OnUpdate(JobHandle inputDeps)
         {
             m_Buffers.Clear();
-            var chunkArray = m_OverlayCanvasQuery.CreateArchetypeChunkArray(Allocator.TempJob);
-            var containerType = GetArchetypeChunkSharedComponentType<CanvasCommandBufferContainer>();
-            for (int i = 0; i < chunkArray.Length; i++)
+            NativeArray<CanvasLayer> layerEntity;
+            var canvasLayerIdFromEntity = GetComponentDataFromEntity<CanvasSortLayer>(true);
+            using (var roots = m_ScreenSpaceOverlayQuery.ToEntityArray(Allocator.TempJob))
             {
-                var commandBuffer = chunkArray[i].GetSharedComponentData(containerType, EntityManager);
-                m_Buffers.Add(commandBuffer.Value);
+                layerEntity = new NativeArray<CanvasLayer>(roots.Length, Allocator.TempJob);
+                for (int i = 0; i < layerEntity.Length; i++)
+                    layerEntity[i] = new CanvasLayer()
+                    {
+                        CanvasEntity = roots[i],
+                        SortId = canvasLayerIdFromEntity[roots[i]].Value
+                    };
+                layerEntity.Sort();
             }
-            chunkArray.Dispose();
+            for (int i = 0; i < layerEntity.Length; i++)
+            {
+                if (EntityManager.HasComponent(layerEntity[i].CanvasEntity,
+                    typeof(CanvasCommandBufferContainer)))
+                {
+                    var commandBuffer = EntityManager.GetSharedComponentData<CanvasCommandBufferContainer>(layerEntity[i].CanvasEntity);
+                    m_Buffers.Add(commandBuffer.Value);
+                }
+            }
+            layerEntity.Dispose();
+            EntityManager.AddComponent(m_UnitializedCanvasQuery, typeof(OverlayCanvasInitialized));
+            EntityManager.RemoveComponent(m_DestroyedCanvasQuery, typeof(OverlayCanvasInitialized));
             return inputDeps;
         }
     }